[#21] Small refactor in the tests
We have to call `JSON.parse` and `JSON.stringify` to be able to compare what
we've got from the API (a JSON) with what comes from the model. We can't
compare them directly because objects like Dates are stringified when converted
to JSON, but not parsed back automatically when loading the JSON.

Stringifying and parsing the output from out internal code makes the results
comparable.