[#65] Return at most 100 pages
There're performance issues on paginating too many results. GitHub limits
1,000 results per query, for example. I would like to do the same, but that
would require validating on two query parameters (`page` and `per_page`). It's
easier to simply limit on a single parameter, so that's what I've done here. If
we have performance issues, we can improve on this solution.

In the worst case, a user would be able to get 100 pages with 100 results each,
totaling 10,000 results.

opentrials/opentrials#65