[#55] Fix autocomplete indexing to work on ES2 and ES1
We shouldn't need this, as we only use ElasticSearch 1.5.2 in staging and
production. However, for the apps automatically deployed for pull requests, we
use the SearchBox plugin. We can set a different ElasticSearch version in
SearchBox by using the `--es_version` option, but there's currently no way to
add plugins' options to Heroku's app.json.

This change should keep the same behaviour both in ES2 and ES1, so it's OK. In
the long run, though, we need a way to specify ElasticSearch's versions (and
Postgres' as well).