[#58] Add Record model
A Record contains the non-deduplicated Trial attributes taken from the Source.
A Trial, in this sense, is generated from its many Records. This is useful to
allow the user to check the original trial's data.

This is the first entity which serves as an relationship table for the
many-to-many relationship between Trial and Source, while also having many
important columns on its own (and not simply "role", as is the case of the
other many relationship tables). This required a few changes to the code and
made the API responses a bit inconsistent.

For the other relationship tables, we have an "attributes" field, but not this
one. The idea of the "attributes" was to avoid naming clashes between the
attributes in the relationship table and in the entity itself. However, it
might be overkill. Maybe it would be better to simply avoid these clashes by
taking care of our namings. I'll make this changes in a later commit.

opentrials/opentrials#58