Add end to end tests
These tests don't hit an actual OpenTrials API instance, but a local server. It
shouldn't make a difference, but it's useful to have a test suite where you can
point a specific server and it would check if it's working or not.

The meaning of e2e tests isn't consistent with OpenTrials itself as well. We
have to fix this later.