[#152] Expose trials and records identifiers
They live in the `identifiers` attribute, which has a format like:

`{ "nct": "NCT0000000", "isrctn": "ISRCTN0000000" }`

The dict keys are Source's IDs, and their values are the Trial's identifier in
that source. The primary source is the one with key equal to
`trial.primary_source_id`.

The same pattern happens in Records as well.