Add API docs in /v1/docs endpoint (#64)
This is generated using Swagger UI. Unfortunately, there's no way to configure
the swagger file's path in Swagger UI, so we need to directly change it in the
code. To avoid doing this change manually, I've added a small bit of code that
does it when starting the server. It's ugly and hacky, but it works for now.