[#254] Returns documents and change relationship between it and trials
We had a `trials_documents` table to generate a many-to-many relationship
between trials and documents. However, a single document can't cite multiple
trials (AFAIK), so we can simplify that to a 1-to-many relationship. This means
we can remove the `trials_documents` table.

I also removed the `documents.slug` column, as we're not using it, and added
the `type` (e.g. csr, epar_segment, ...) and `url` columns. For now, the trials
endpoint returns an array of these documents with their `name`, `type` and
`url`.

opentrials/opentrials#254