[#302] Add fda_approvals table with a relation to documents
This table contains the approval history for the FDA applications. A document
now can be related to either a trial or a FDA Approval (but not both).
There're many ways of modeling this relationship (e.g.  polymorphic
association, creating a new association table, etc.). Here I implemented the
simplest one. If, in the future, documents can be related to more entity types,
we can refactor the database in another structure.

A document MUST have either a trial_id or an intervention_id (but not
both).

opentrials/opentrials#302