[#364] Reduce memory usage of the reindexing process
I did so by offloading the removal of entities without trials to the database,
instead of doing it with NodeJS. With these changes, there's no need to query
for the actual trials as well, so it reduces the number of SQL queries in half.

I haven't tested the difference in memory usage, but considering we're both not
loading the trials in memory and loading just the entities we need, I'd imagine
it to be quite significant. Hopefully enough to avoid the Out of Memory errors
we're seeing.

Fixes opentrials/opentrials#364