[#387] Add entity File that belongs to many Documents
This extracts the attributes related to the actual file location and its text
from the `documents` table to a new `files` table. The reasoning behind it is
that multiple documents can (and do) reference the same file, so this
extraction will allow us to keep a single entry in our database in those cases.

I haven't removed the columns from the `documents` table yet because I want to
populate the `files` table with our current contents before that. After we run
this migration and populate the `files` table, we can add another migration to
remove the then redudant columns.

I also changed to use the "documentcloud_id" instead of its URL so we can
delete files from DocumentCloud as needed (e.g. when they're updated). This
isn't ideal, but we can generate a document's URL based on its ID.

The API changes in this pull request are needed just while we don't finish the
data migration. They will be reverted afterwards.

opentrials/opentrials#387