Add uniqueness constraints to documents table
There are 3 cases for a document. It can be an FDA document, a document with a
file (with our without URL), and a document with a URL without a file. The
uniqueness constraints for them are:

1. FDA documents - [`fda_approval_id`, `file_id`, `name`]
2. Documents with files - [`type`, `file_id`]
3. Documents with URLs - [`type`, `url`]

In this order. I mean, if a document has an `fda_approval_id`, its the checked
uniqueness constraint must be `1`, and so on.

This is getting overly complicated, and might be a smell that we're trying to
cram different entities in the same table. We might extract them in the future.