[#500] Change Document's and DocumentSummary responses
The `/v1/documents` endpoint now returns document's summaries, instead of the
actual documents. The main reason for this is to avoid returning the actual
file's pages in that endpoint, which can be very big.

In the future, we'll probably add an option to hide the file's pages even in
the `/v1/documents/<id>`, allowing the user to get smaller responses.

opentrials/opentrials#500