Catch all errors in the controllers' promises
This is done simply by moving the `.catch()` to be the last call in the Promise
chain. This means that any error in any of the previous commands will be
caught.

If we don't do this and an uncaught error happens, the Promise won't stop and
we'll get a timeout.