[#500] Create /v1/search/fda_documents endpoint
The main requirements for this search endpoint are:

* User can search inside the documents files pages' text;
* It returns only the matching pages, so we can tell the user where the text
  was found

To return only the matching pages, we can't use a single index type (nested
mappings might work, but not without its own caveats). So we create two index
types: `document` and `page`. The `page` type is of the format:

```
{
  text: 'Lorem ipsum',
  num: 1
}
```

Where `text` is the content of that page, and `num` is the page number. The
`page` type is set as a child of its related `document`. This way we can query
for documents which have pages with a specific text, and can return both the
`document` and the `page` in a single query by using the `inner_hits` query.

There's one known problem with this: ElasticSearch only allows a 1-to-many
relationship between parents and children. In other words, a child can have
only one parent. However, in our database, multiple documents can point to the
same file. About 15% of the FDA documents point to the same file. This means
we're reindexing the same file multiple times, which is obviously wasteful.

The only solution I've found for this was doing multiple search queries and
joining them on the client-side (i.e. our API code). That felt more
complicated, so I didn't want to go that route for now (deadlines, you
know...). This problem is just something we should be aware of, and deal with
it *if* it becomes a problem.

opentrials/opentrials#500