[#693] Fix issue when reindexing trials
We were indexing Bookshelf models, instead of simple JSON objects. For some
reason, this was causing ElasticSearch to complain with a
`illegal_argument_exception` in some data.

I'm not sure what exact was the culprit, but with this change we don't get the
exception anymore. As production is currently down because of it, I'm moving on
even though I don't fully understand the cause of the exception.

opentrials/opentrials#693