Remove tables that already have default timestamps from migration
If we keep them in the migration, it means that when that migrations is rolled
back, it'll remove the defaults updated_at/created_at from tables that used to
have (e.g. `fda_applications`). This way we break the assumption that rolling
back a migration must return the DB to its previous state.